+++ exam-ning-springcloud-system-user/src/main/java/com/ning/infrastructure/persistence/repository/impl/UserRepositoryImpl.java
@@ -160,7 +160,7 @@ public class UserRepositoryImpl implements UserRepository {
         }
 
         UserDO userDO = userDOOpt.get();
-        userDO.setIsDeleted(1);
+        userDO.setIsDeleted((byte) 1);
         userDao.updateById(userDO);
         return true;
     }
